requirements.txt
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
bitsandbytes==0.48.1
brotli==1.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
cryptography==46.0.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
emoji==2.15.0
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.119.0
ffmpy==0.6.4
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.8.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
inquirerpy==0.3.4
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jiwer==4.0.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.9.1
langchain==0.3.27
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-experimental==0.3.4
langchain-google-genai==3.0.0
langchain-huggingface==1.0.0
langchain-ollama==1.0.0
langchain-openai==1.0.0
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.38
llama_cpp_python==0.3.16
llm_mapreduce==0.1.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
nvidia-cublas==13.0.0.19
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti==13.0.48
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime==13.0.48
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver==12.0.3.29
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse==12.6.2.49
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-cusparselt-cu13==0.8.0
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvidia-nvtx-cu12==12.4.127
olefile==0.47
ollama==0.6.0
omegaconf==2.3.0
onnxruntime==1.23.1
onnxruntime-gpu==1.23.0
openai==2.5.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.0
ptyprocess==0.7.0
punctuators==0.0.7
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pynini==2.1.7
pypdf==6.1.3
pytest==9.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rank-bm25==0.2.2
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
ruff==0.14.4
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.1
tenacity==9.1.2
TextGrid==1.6.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.2.0
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
